cpp11-package {cpp11}	R Documentation

cpp11: A C++11 Interface for R's C Interface

Description

Provides a header only, C++11 interface to R's C interface. Compared to other approaches 'cpp11' strives to be safe against long jumps from the C API as well as C++ exceptions, conform to normal R function semantics and supports interaction with 'ALTREP' vectors.

Author(s)

Maintainer: Romain François romain@rstudio.com

Authors:

Jim Hester (ORCID)

Other contributors:

Benjamin Kietzman [contributor]

RStudio [copyright holder, funder]

See Also

Useful links:

https://cpp11.r-lib.org

https://github.com/r-lib/cpp11

Report bugs at https://github.com/r-lib/cpp11/issues

[Package cpp11 version 0.4.2 Index]