Compile C++ code

cpp_function() is a helper which makes it easier to compile a single function for use in R. eval_cpp() is a helper to evaluate a single expression and return the result.

source_cpp(file, code = NULL, env = parent.frame(), clean = TRUE, quiet = TRUE)

cpp_function(code, env = parent.frame())

eval_cpp(code)

Arguments

file
A file containing C++ code to compile

code
If non-null, the C++ code to compile

env
The R environment where the R wrapping functions should be defined.

clean
If TRUE, cleanup the files after sourcing

quiet
If 'TRUE`, do not show compiler output

Examples

if (FALSE) {
source_cpp(
  code = '#include "cpp11/integers.hpp"

  [[cpp11::export]]
  int num_odd(cpp11::integers x) {
    int total = 0;
    for (int val : x) {
      if ((val % 2) == 1) {
        ++total;
      }
    }
    return total;
  }
  ')

num_odd(as.integer(c(1:10, 15, 23)))
}